Rename utils_jax to utils
